requirements.txt
greenlet==3.0.3
chromadb>=0.50.0
hnswlib>=0.8.0
flask==3.0.2
python-dotenv==1.0.1
pyyaml==6.0.1
google-generativeai==0.5.0
sentence-transformers==2.5.1
deep-translator==1.11.4
beautifulsoup4==4.12.3
playwright==1.44.0
requests==2.31.0
langchain-core==0.1.42
langgraph==0.0.35
